Associate user with new posts + verify user before updating or deleting

+++ server/src/routes/posts.js
@@ -2,6 +2,7 @@ const express = require('express');
 const { firebaseAuthMiddleware } = require('../middleware/firebaseAuth');
 
 const router = express.Router();
+const User = require('../db/models/users');
 const Post = require('../db/models/post');
 const Comment = require('../db/models/comments');
 const SubThread = require('../db/models/subThreads');
@@ -18,13 +19,25 @@ router.get('/', async (req, res) => {
 
 // Create one post
 router.post('/', firebaseAuthMiddleware, async (req, res) => {
+  let foundUser = await User.findOne({ email: req.user.email });
+  if (!foundUser) {
+    const user = new User({
+      email: req.user.email,
+    });
+
+    const newUser = await user.save();
+    foundUser = await User.findOne({ email: newUser.email });
+  }
+
   const post = new Post({
     title: req.body.title,
     body: req.body.body,
+    author: foundUser._id,
   });
 
   try {
     const newPost = await post.save();
+    await User.update({ _id: foundUser._id }, { $push: { posts: newPost._id } });
     res.status(201).json(newPost);
   } catch (err) {
     res.status(400).json({ message: err.message });
@@ -34,15 +47,21 @@ router.post('/', firebaseAuthMiddleware, async (req, res) => {
 // Update one post
 // eslint-disable-next-line no-unused-vars
 router.put('/:id', firebaseAuthMiddleware, async (req, res) => {
-  try {
-    if (req.body.title != null && req.body.body != null) {
-      await Post.update({ _id: req.params.id }, { title: req.body.title, body: req.body.body });
-      res.status(200).send();
-    } else {
-      res.status(400).json({ message: 'Please include a title and body' });
+  const currentPost = await Post.findById(req.params.id);
+  const user = await User.findById(currentPost.author);
+  if (user.email === req.user.email) {
+    try {
+      if (req.body.title != null && req.body.body != null) {
+        await Post.update({ _id: req.params.id }, { title: req.body.title, body: req.body.body });
+        res.status(200).send();
+      } else {
+        res.status(400).json({ message: 'Please include a title and body' });
+      }
+    } catch (err) {
+      res.status(404).json({ message: err.message });
     }
-  } catch (err) {
-    res.status(404).json({ message: err.message });
+  } else {
+    res.status(400).json({ message: 'You must be the author to update a post' });
   }
 });
 
@@ -80,13 +99,20 @@ router.get('/:id', async (req, res) => {
 // Deleting one post
 // eslint-disable-next-line no-unused-vars
 router.delete('/:id', firebaseAuthMiddleware, async (req, res) => {
-  try {
-    await Post.findOneAndDelete({
-      _id: req.params.id,
-    });
-    res.status(200).send();
-  } catch (err) {
-    res.status(404).json({ message: err.message });
+  const currentPost = await Post.findById(req.params.id);
+  const user = await User.findById(currentPost.author);
+
+  if (user.email === req.user.email) {
+    try {
+      await Post.findOneAndDelete({
+        _id: req.params.id,
+      });
+      res.status(200).send();
+    } catch (err) {
+      res.status(404).json({ message: err.message });
+    }
+  } else {
+    res.status(400).json({ message: 'You must be the author to delete a post' });
   }
 });
